+++ kapitan/utils.py
@@ -18,7 +18,6 @@ import shutil
 import stat
 import sys
 import tarfile
-import traceback
 from collections import Counter, defaultdict
 from datetime import time
 from functools import lru_cache
@@ -30,9 +29,8 @@ import magic
 import requests
 import yaml
 
-from kapitan import cached, defaults
+from kapitan import cached
 from kapitan.errors import CompileError
-from kapitan.jinja2_filters import load_jinja2_filters, load_jinja2_filters_from_file
 from kapitan.version import VERSION
 
 logger = logging.getLogger(__name__)
@@ -82,81 +80,6 @@ def render_jinja2_template(content, context):
 def sha256_string(string):
     """Returns sha256 hex digest for string"""
     return sha256(string.encode("UTF-8")).hexdigest()
-
-
-def _jinja_error_info(trace_data):
-    """Extract jinja2 templating related frames from traceback data"""
-    try:
-        return [x for x in trace_data if x[2] in ("top-level template code", "template", "<module>")][-1]
-    except IndexError:
-        pass
-
-
-def render_jinja2_file(name, context, jinja2_filters=defaults.DEFAULT_JINJA2_FILTERS_PATH, search_paths=None):
-    """Render jinja2 file name with context"""
-    path, filename = os.path.split(name)
-    search_paths = [path or "./"] + (search_paths or [])
-    env = jinja2.Environment(
-        undefined=jinja2.StrictUndefined,
-        loader=jinja2.FileSystemLoader(search_paths),
-        trim_blocks=True,
-        lstrip_blocks=True,
-        extensions=["jinja2.ext.do"],
-    )
-    load_jinja2_filters(env)
-    load_jinja2_filters_from_file(env, jinja2_filters)
-    try:
-        return env.get_template(filename).render(context)
-    except jinja2.TemplateError as e:
-        # Exception misses the line number info. Retreive it from traceback
-        err_info = _jinja_error_info(traceback.extract_tb(sys.exc_info()[2]))
-        raise CompileError(f"Jinja2 TemplateError: {e}, at {err_info[0]}:{err_info[1]}")
-
-
-def render_jinja2(path, context, jinja2_filters=defaults.DEFAULT_JINJA2_FILTERS_PATH, search_paths=None):
-    """
-    Render files in path with context
-    Returns a dict where the is key is the filename (with subpath)
-    and value is a dict with content and mode
-    Empty paths will not be rendered
-    Path can be a single file or directory
-    Ignores hidden files (.filename)
-    """
-    rendered = {}
-    walk_root_files = []
-    if os.path.isfile(path):
-        dirname = os.path.dirname(path)
-        basename = os.path.basename(path)
-        walk_root_files = [(dirname, None, [basename])]
-    else:
-        walk_root_files = os.walk(path)
-
-    for root, _, files in walk_root_files:
-        for f in files:
-            if f.startswith("."):
-                logger.debug("render_jinja2: ignoring file %s", f)
-                continue
-            render_path = os.path.join(root, f)
-            logger.debug("render_jinja2 rendering %s", render_path)
-            # get subpath and filename, strip any leading/trailing /
-            name = render_path[len(os.path.commonprefix([root, path])) :].strip("/")
-            try:
-                rendered[name] = {
-                    "content": render_jinja2_file(
-                        render_path, context, jinja2_filters=jinja2_filters, search_paths=search_paths
-                    ),
-                    "mode": file_mode(render_path),
-                }
-            except Exception as e:
-                raise CompileError(f"Jinja2 error: failed to render {render_path}: {e}")
-
-    return rendered
-
-
-def file_mode(name):
-    """Returns mode for file name"""
-    st = os.stat(name)
-    return stat.S_IMODE(st.st_mode)
 
 
 def prune_empty(d):
